+++ .anvil_editor.yaml
@@ -12,6 +12,7 @@ unique_ids:
   modules:
     forms: '1731061133551283557447632.6057'
     forms.components: '1731068503683675445652055.4076'
+    forms.helpers: '1740752268044583436869527.1309'
     forms.layouts: '1731061141288775989625445.2242'
     forms.pages: '1731061151926175732205682.1231'
     forms.pages.author: '1731063111076555002607775.9423'
